Refactored composition logic into discrete priority-ordered rules to allow for extensibility.
